+++ pkg/models/alerts.go
@@ -5,23 +5,23 @@ import (
 )
 
 type AlertRule struct {
-	Id           int64  `json:"id"`
-	OrgId        int64  `json:"-"`
-	DatasourceId int64  `json:"datasourceId"`
-	DashboardId  int64  `json:"dashboardId"`
-	PanelId      int64  `json:"panelId"`
-	Query        string `json:"query"`
-	QueryRefId   string `json:"queryRefId"`
-	WarnLevel    int64  `json:"warnLevel"`
-	CritLevel    int64  `json:"critLevel"`
-	WarnOperator string `json:"warnOperator"`
-	CritOperator string `json:"critOperator"`
-	Frequency    int64  `json:"frequency"`
-	Title        string `json:"title"`
-	Description  string `json:"description"`
-	QueryRange   int    `json:"queryRange"`
-	Aggregator   string `json:"aggregator"`
-	State        string `json:"state"`
+	Id           int64   `json:"id"`
+	OrgId        int64   `json:"-"`
+	DatasourceId int64   `json:"datasourceId"`
+	DashboardId  int64   `json:"dashboardId"`
+	PanelId      int64   `json:"panelId"`
+	Query        string  `json:"query"`
+	QueryRefId   string  `json:"queryRefId"`
+	WarnLevel    float64 `json:"warnLevel"`
+	CritLevel    float64 `json:"critLevel"`
+	WarnOperator string  `json:"warnOperator"`
+	CritOperator string  `json:"critOperator"`
+	Frequency    int64   `json:"frequency"`
+	Title        string  `json:"title"`
+	Description  string  `json:"description"`
+	QueryRange   int     `json:"queryRange"`
+	Aggregator   string  `json:"aggregator"`
+	State        string  `json:"state"`
 
 	Created time.Time `json:"created"`
 	Updated time.Time `json:"updated"`
